roadmap_requirements.txt
Django==1.3.0
South==0.7.3
Whoosh==2.1.0
django-avatar==1.0.5
django-flash==1.8
django-haystack==1.2.4
django-reversion==1.4
django-tagging==0.3.1
django-tinymce==1.5.1a2
simplejson==2.0.9
wsgiref==0.1.2
#MySQL-python==1.2.3
PIL==1.1.7
fabric==1.4
